refactor organization member dialogs
